pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain[openai]>=1.1.2",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
